+++ client-sdks/client-js/src/types.ts
@@ -1,5 +1,5 @@
 import type {
-  MessageType,
+  MastraMessageV1,
   AiMessageType,
   CoreMessage,
   QueryResult,
@@ -7,10 +7,10 @@ import type {
   WorkflowRuns,
   LegacyWorkflowRuns,
 } from '@mastra/core';
+import type { AgentGenerateOptions, AgentStreamOptions } from '@mastra/core/agent';
 import type { BaseLogMessage } from '@mastra/core/logger';
 
-import type { AgentGenerateOptions, AgentStreamOptions } from '@mastra/core/agent';
-import type { ServerInfo } from '@mastra/core/mcp';
+import type { MCPToolType, ServerInfo } from '@mastra/core/mcp';
 import type { RuntimeContext } from '@mastra/core/runtime-context';
 import type { Workflow, WatchEvent, WorkflowResult } from '@mastra/core/workflows';
 import type {
@@ -60,6 +60,8 @@ export interface GetAgentResponse {
   workflows: Record<string, GetWorkflowResponse>;
   provider: string;
   modelId: string;
+  defaultGenerateOptions: WithoutMethods<AgentGenerateOptions>;
+  defaultStreamOptions: WithoutMethods<AgentStreamOptions>;
 }
 
 export type GenerateParams<T extends JSONSchema7 | ZodSchema | undefined = undefined> = {
@@ -170,11 +172,11 @@ export interface GetVectorIndexResponse {
 }
 
 export interface SaveMessageToMemoryParams {
-  messages: MessageType[];
+  messages: MastraMessageV1[];
   agentId: string;
 }
 
-export type SaveMessageToMemoryResponse = MessageType[];
+export type SaveMessageToMemoryResponse = MastraMessageV1[];
 
 export interface CreateMemoryThreadParams {
   title?: string;
@@ -304,6 +306,7 @@ export interface McpToolInfo {
   name: string;
   description?: string;
   inputSchema: string;
+  toolType?: MCPToolType;
 }
 
 export interface McpServerToolListResponse {