+++ packages/core/src/workflows/default.ts
@@ -1,11 +1,12 @@
 import { context as otlpContext, trace } from '@opentelemetry/api';
 import type { Span } from '@opentelemetry/api';
 import type { RuntimeContext } from '../di';
+import { EMITTER_SYMBOL } from './constants';
 import type { ExecutionGraph } from './execution-engine';
 import { ExecutionEngine } from './execution-engine';
 import type { ExecuteFunction, Step } from './step';
-import type { StepResult } from './types';
-import type { StepFlowEntry } from './workflow';
+import type { StepResult, StepSuccess } from './types';
+import type { SerializedStepFlowEntry, StepFlowEntry } from './workflow';
 
 export type ExecutionContext = {
   workflowId: string;
@@ -26,8 +27,8 @@ export class DefaultExecutionEngine extends ExecutionEngine {
   protected async fmtReturnValue<TOutput>(
     executionSpan: Span | undefined,
     emitter: { emit: (event: string, data: any) => Promise<void> },
-    stepResults: Record<string, StepResult<any>>,
-    lastOutput: StepResult<any>,
+    stepResults: Record<string, StepResult<any, any, any, any>>,
+    lastOutput: StepResult<any, any, any, any>,
     error?: Error | string,
   ): Promise<TOutput> {
     const base: any = {
@@ -62,11 +63,17 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         eventTimestamp: Date.now(),
       });
 
-      base.error = error instanceof Error ? error : (lastOutput.error ?? new Error('Unknown error: ' + error));
+      base.error =
+        error instanceof Error
+          ? (error?.stack ?? error)
+          : (lastOutput.error ??
+            (typeof error === 'string'
+              ? error
+              : (new Error('Unknown error: ' + error)?.stack ?? new Error('Unknown error: ' + error))));
     } else if (lastOutput.status === 'suspended') {
       const suspendedStepIds = Object.entries(stepResults).flatMap(([stepId, stepResult]) => {
         if (stepResult?.status === 'suspended') {
-          const nestedPath = stepResult?.payload?.__workflow_meta?.path;
+          const nestedPath = stepResult?.suspendPayload?.__workflow_meta?.path;
           return nestedPath ? [[stepId, ...nestedPath]] : [[stepId]];
         }
 
@@ -102,11 +109,12 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     workflowId: string;
     runId: string;
     graph: ExecutionGraph;
+    serializedStepGraph: SerializedStepFlowEntry[];
     input?: TInput;
     resume?: {
       // TODO: add execute path
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
@@ -129,8 +137,6 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       attributes: { componentName: workflowId, runId },
     });
 
-    await this.mastra?.getStorage()?.init();
-
     let startIdx = 0;
     if (resume?.resumePath) {
       startIdx = resume.resumePath[0]!;
@@ -146,6 +152,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           workflowId,
           runId,
           entry,
+          serializedStepGraph: params.serializedStepGraph,
           prevStep: steps[i - 1]!,
           stepResults,
           resume,
@@ -215,7 +222,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     workflowId: string;
     runId: string;
     step: Step<string, any, any>;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     executionContext: ExecutionContext;
     resume?: {
       steps: string[];
@@ -224,13 +231,25 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     prevOutput: any;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
+    const startTime = resume?.steps[0] === step.id ? undefined : Date.now();
+    const resumeTime = resume?.steps[0] === step.id ? Date.now() : undefined;
+
+    const stepInfo = {
+      ...stepResults[step.id],
+      payload: prevOutput,
+      ...(resume?.steps[0] === step.id ? { resumePayload: resume?.resumePayload } : {}),
+      ...(startTime ? { startedAt: startTime } : {}),
+      ...(resumeTime ? { resumedAt: resumeTime } : {}),
+    };
+
     await emitter.emit('watch', {
       type: 'watch',
       payload: {
         currentStep: {
           id: step.id,
           status: 'running',
+          ...stepInfo,
         },
         workflowState: {
           status: 'running',
@@ -238,6 +257,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             ...stepResults,
             [step.id]: {
               status: 'running',
+              ...stepInfo,
             },
           },
           result: null,
@@ -245,6 +265,12 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         },
       },
       eventTimestamp: Date.now(),
+    });
+    await emitter.emit('watch-v2', {
+      type: 'step-start',
+      payload: {
+        id: step.id,
+      },
     });
 
     const _runStep = (step: Step<any, any, any, any>, spanName: string, attributes?: Record<string, string>) => {
@@ -278,6 +304,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       try {
         let suspended: { payload: any } | undefined;
         const result = await runStep({
+          runId,
           mastra: this.mastra!,
           runtimeContext,
           inputData: prevOutput,
@@ -303,21 +330,30 @@ export class DefaultExecutionEngine extends ExecutionEngine {
             steps: resume?.steps?.slice(1) || [],
             resumePayload: resume?.resumePayload,
             // @ts-ignore
-            runId: stepResults[step.id]?.payload?.__workflow_meta?.runId,
+            runId: stepResults[step.id]?.suspendPayload?.__workflow_meta?.runId,
           },
-          emitter,
+          [EMITTER_SYMBOL]: emitter,
         });
 
         if (suspended) {
-          execResults = { status: 'suspended', payload: suspended.payload };
+          execResults = { status: 'suspended', suspendPayload: suspended.payload, suspendedAt: Date.now() };
         } else {
-          execResults = { status: 'success', output: result };
+          execResults = { status: 'success', output: result, endedAt: Date.now() };
         }
 
         break;
       } catch (e) {
         this.logger.error('Error executing step: ' + ((e as Error)?.stack ?? e));
-        execResults = { status: 'failed', error: e instanceof Error ? e : new Error('Unknown error: ' + e) };
+        execResults = {
+          status: 'failed',
+          error:
+            e instanceof Error
+              ? (e?.stack ?? e)
+              : typeof e === 'string'
+                ? e
+                : (new Error('Unknown error: ' + e)?.stack ?? new Error('Unknown error: ' + e)),
+          endedAt: Date.now(),
+        };
       }
     }
 
@@ -326,18 +362,16 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       payload: {
         currentStep: {
           id: step.id,
-          status: execResults.status,
-          output: execResults.output,
+          ...stepInfo,
+          ...execResults,
         },
         workflowState: {
           status: 'running',
           steps: {
             ...stepResults,
             [step.id]: {
-              status: execResults.status,
-              output: execResults.output,
-              error: execResults.error,
-              payload: execResults.payload,
+              ...stepInfo,
+              ...execResults,
             },
           },
 
@@ -348,7 +382,34 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       eventTimestamp: Date.now(),
     });
 
-    return execResults;
+    if (execResults.status === 'suspended') {
+      await emitter.emit('watch-v2', {
+        type: 'step-suspended',
+        payload: {
+          id: step.id,
+          output: execResults.output,
+        },
+      });
+    } else {
+      await emitter.emit('watch-v2', {
+        type: 'step-result',
+        payload: {
+          id: step.id,
+          status: execResults.status,
+          output: execResults.output,
+        },
+      });
+
+      await emitter.emit('watch-v2', {
+        type: 'step-finish',
+        payload: {
+          id: step.id,
+          metadata: {},
+        },
+      });
+    }
+
+    return { ...stepInfo, ...execResults };
   }
 
   async executeParallel({
@@ -356,6 +417,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     runId,
     entry,
     prevStep,
+    serializedStepGraph,
     stepResults,
     resume,
     executionContext,
@@ -365,20 +427,21 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     workflowId: string;
     runId: string;
     entry: { type: 'parallel'; steps: StepFlowEntry[] };
+    serializedStepGraph: SerializedStepFlowEntry[];
     prevStep: StepFlowEntry;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     resume?: {
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
     executionContext: ExecutionContext;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
     let execResults: any;
-    const results: StepResult<any>[] = await Promise.all(
+    const results: StepResult<any, any, any, any>[] = await Promise.all(
       entry.steps.map((step, i) =>
         this.executeEntry({
           workflowId,
@@ -386,6 +449,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           entry: step,
           prevStep,
           stepResults,
+          serializedStepGraph,
           resume,
           executionContext: {
             workflowId,
@@ -405,7 +469,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     if (hasFailed) {
       execResults = { status: 'failed', error: hasFailed.error };
     } else if (hasSuspended) {
-      execResults = { status: 'suspended', payload: hasSuspended.payload };
+      execResults = { status: 'suspended', payload: hasSuspended.suspendPayload };
     } else {
       execResults = {
         status: 'success',
@@ -429,6 +493,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     entry,
     prevOutput,
     prevStep,
+    serializedStepGraph,
     stepResults,
     resume,
     executionContext,
@@ -437,26 +502,28 @@ export class DefaultExecutionEngine extends ExecutionEngine {
   }: {
     workflowId: string;
     runId: string;
+    serializedStepGraph: SerializedStepFlowEntry[];
     entry: { type: 'conditional'; steps: StepFlowEntry[]; conditions: ExecuteFunction<any, any, any, any>[] };
     prevStep: StepFlowEntry;
     prevOutput: any;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     resume?: {
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
     executionContext: ExecutionContext;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
     let execResults: any;
     const truthyIndexes = (
       await Promise.all(
         entry.conditions.map(async (cond, index) => {
           try {
             const result = await cond({
+              runId,
               mastra: this.mastra!,
               runtimeContext,
               inputData: prevOutput,
@@ -476,7 +543,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
 
               // TODO: this function shouldn't have suspend probably?
               suspend: async (_suspendPayload: any) => {},
-              emitter,
+              [EMITTER_SYMBOL]: emitter,
             });
             return result ? index : null;
           } catch (e: unknown) {
@@ -488,7 +555,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     ).filter((index): index is number => index !== null);
 
     const stepsToRun = entry.steps.filter((_, index) => truthyIndexes.includes(index));
-    const results: StepResult<any>[] = await Promise.all(
+    const results: StepResult<any, any, any, any>[] = await Promise.all(
       stepsToRun.map((step, index) =>
         this.executeEntry({
           workflowId,
@@ -496,6 +563,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           entry: step,
           prevStep,
           stepResults,
+          serializedStepGraph,
           resume,
           executionContext: {
             workflowId,
@@ -515,7 +583,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     if (hasFailed) {
       execResults = { status: 'failed', error: hasFailed.error };
     } else if (hasSuspended) {
-      execResults = { status: 'suspended', payload: hasSuspended.payload };
+      execResults = { status: 'suspended', payload: hasSuspended.suspendPayload };
     } else {
       execResults = {
         status: 'success',
@@ -554,20 +622,20 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     };
     prevStep: StepFlowEntry;
     prevOutput: any;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     resume?: {
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
     executionContext: ExecutionContext;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
     const { step, condition } = entry;
     let isTrue = true;
-    let result: StepResult<any> = { status: 'success', output: prevOutput };
+    let result = { status: 'success', output: prevOutput } as unknown as StepResult<any, any, any, any>;
 
     do {
       result = await this.executeStep({
@@ -577,7 +645,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         stepResults,
         executionContext,
         resume,
-        prevOutput: result.output,
+        prevOutput: (result as { output: any }).output,
         emitter,
         runtimeContext,
       });
@@ -587,6 +655,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       }
 
       isTrue = await condition({
+        runId,
         mastra: this.mastra!,
         runtimeContext,
         inputData: result.output,
@@ -600,7 +669,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
           return result?.status === 'success' ? result.output : null;
         },
         suspend: async (_suspendPayload: any) => {},
-        emitter,
+        [EMITTER_SYMBOL]: emitter,
       });
     } while (entry.loopType === 'dowhile' ? isTrue : !isTrue);
 
@@ -629,20 +698,22 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     };
     prevStep: StepFlowEntry;
     prevOutput: any;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     resume?: {
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
     executionContext: ExecutionContext;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
     const { step, opts } = entry;
-    const results: StepResult<any>[] = [];
+    const results: StepResult<any, any, any, any>[] = [];
     const concurrency = opts.concurrency;
+    const startTime = resume?.steps[0] === step.id ? undefined : Date.now();
+    const resumeTime = resume?.steps[0] === step.id ? Date.now() : undefined;
 
     for (let i = 0; i < prevOutput.length; i += concurrency) {
       const items = prevOutput.slice(i, i + concurrency);
@@ -671,18 +742,30 @@ export class DefaultExecutionEngine extends ExecutionEngine {
       }
     }
 
-    return { status: 'success', output: results };
+    return {
+      ...stepResults[step.id],
+      status: 'success',
+      payload: prevOutput,
+      ...(resume?.steps[0] === step.id ? { resumePayload: resume?.resumePayload } : {}),
+      output: results,
+      //@ts-ignore
+      endedAt: Date.now(),
+      ...(startTime ? { startedAt: startTime } : {}),
+      ...(resumeTime ? { resumedAt: resumeTime } : {}),
+    } as StepSuccess<any, any, any, any>;
   }
 
   protected async persistStepUpdate({
     workflowId,
     runId,
     stepResults,
+    serializedStepGraph,
     executionContext,
   }: {
     workflowId: string;
     runId: string;
-    stepResults: Record<string, StepResult<any>>;
+    stepResults: Record<string, StepResult<any, any, any, any>>;
+    serializedStepGraph: SerializedStepFlowEntry[];
     executionContext: ExecutionContext;
   }) {
     await this.mastra?.getStorage()?.persistWorkflowSnapshot({
@@ -693,6 +776,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         value: {},
         context: stepResults as any,
         activePaths: [],
+        serializedStepGraph,
         suspendedPaths: executionContext.suspendedPaths,
         // @ts-ignore
         timestamp: Date.now(),
@@ -705,6 +789,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     runId,
     entry,
     prevStep,
+    serializedStepGraph,
     stepResults,
     resume,
     executionContext,
@@ -715,17 +800,18 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     runId: string;
     entry: StepFlowEntry;
     prevStep: StepFlowEntry;
-    stepResults: Record<string, StepResult<any>>;
+    serializedStepGraph: SerializedStepFlowEntry[];
+    stepResults: Record<string, StepResult<any, any, any, any>>;
     resume?: {
       steps: string[];
-      stepResults: Record<string, StepResult<any>>;
+      stepResults: Record<string, StepResult<any, any, any, any>>;
       resumePayload: any;
       resumePath: number[];
     };
     executionContext: ExecutionContext;
     emitter: { emit: (event: string, data: any) => Promise<void> };
     runtimeContext: RuntimeContext;
-  }): Promise<StepResult<any>> {
+  }): Promise<StepResult<any, any, any, any>> {
     const prevOutput = this.getStepOutput(stepResults, prevStep);
     let execResults: any;
 
@@ -749,6 +835,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         runId,
         entry: entry.steps[idx!]!,
         prevStep,
+        serializedStepGraph,
         stepResults,
         resume,
         executionContext: {
@@ -769,6 +856,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         entry,
         prevStep,
         stepResults,
+        serializedStepGraph,
         resume,
         executionContext,
         emitter,
@@ -782,6 +870,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
         prevStep,
         prevOutput,
         stepResults,
+        serializedStepGraph,
         resume,
         executionContext,
         emitter,
@@ -822,6 +911,7 @@ export class DefaultExecutionEngine extends ExecutionEngine {
     await this.persistStepUpdate({
       workflowId,
       runId,
+      serializedStepGraph,
       stepResults,
       executionContext,
     });