+++ examples/dane/CHANGELOG.md
@@ -1,5 +1,102 @@
 # @mastra/dane
 
+## 0.0.23-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [ab5adbe]
+- Updated dependencies [195c428]
+- Updated dependencies [f73e11b]
+  - @mastra/core@0.10.2-alpha.3
+
+## 0.0.23-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [1e8bb40]
+- Updated dependencies [f0d559f]
+  - @mastra/core@0.10.2-alpha.2
+  - @mastra/github@1.2.1-alpha.0
+  - @mastra/memory@0.10.2-alpha.1
+  - @mastra/upstash@0.10.2-alpha.1
+  - @mastra/libsql@0.10.1-alpha.1
+  - @mastra/mcp@0.10.2-alpha.1
+  - @mastra/rag@0.10.2-alpha.0
+
+## 0.0.23-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [ee77e78]
+- Updated dependencies [2901125]
+  - @mastra/core@0.10.2-alpha.1
+
+## 0.0.23-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [592a2db]
+- Updated dependencies [e5dc18d]
+- Updated dependencies [9e37877]
+  - @mastra/core@0.10.2-alpha.0
+  - @mastra/mcp@0.10.2-alpha.0
+  - @mastra/memory@0.10.2-alpha.0
+  - @mastra/upstash@0.10.2-alpha.0
+  - @mastra/libsql@0.10.1-alpha.0
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [d70b807]
+- Updated dependencies [6d16390]
+- Updated dependencies [1e4a421]
+- Updated dependencies [e60402a]
+- Updated dependencies [ab89d6a]
+- Updated dependencies [200d0da]
+- Updated dependencies [bf5f17b]
+- Updated dependencies [5343f93]
+- Updated dependencies [38aee50]
+- Updated dependencies [8784cef]
+- Updated dependencies [5c41100]
+- Updated dependencies [d6a759b]
+- Updated dependencies [f56fd29]
+- Updated dependencies [6015bdf]
+  - @mastra/memory@0.10.1
+  - @mastra/core@0.10.1
+  - @mastra/upstash@0.10.1
+  - @mastra/mcp@0.10.1
+  - @mastra/rag@0.10.1
+
+## 0.0.22-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [d70b807]
+  - @mastra/memory@0.10.1-alpha.0
+  - @mastra/core@0.10.1-alpha.3
+
+## 0.0.22-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [6015bdf]
+  - @mastra/core@0.10.1-alpha.2
+
+## 0.0.22-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [ab89d6a]
+- Updated dependencies [200d0da]
+- Updated dependencies [bf5f17b]
+- Updated dependencies [5343f93]
+- Updated dependencies [38aee50]
+- Updated dependencies [5c41100]
+- Updated dependencies [d6a759b]
+  - @mastra/mcp@0.10.1-alpha.0
+  - @mastra/core@0.10.1-alpha.1
+
 ## 0.0.22-alpha.1
 
 ### Patch Changes